scaling in double

+++ bundles/binding/org.openhab.binding.sapp/src/main/java/org/openhab/binding/sapp/internal/model/SappAddressDecimal.java
@@ -53,8 +53,8 @@ public class SappAddressDecimal extends SappAddress {
 		}
 	}
 	
-	public int scaledValue(int value) {
-		return ((value - originalMinScale) * (maxScale - minScale) / (originalMaxScale - originalMinScale)) + minScale;
+	public double scaledValue(int value) {
+		return (((double) (value - originalMinScale)) * ((double) (maxScale - minScale)) / ((double) (originalMaxScale - originalMinScale))) + ((double) minScale);
 	}
 
 	@Override
